+++ clarifai/client/compute_cluster.py
@@ -10,7 +10,7 @@ from clarifai.client.base import BaseClient
 from clarifai.client.lister import Lister
 from clarifai.client.nodepool import Nodepool
 from clarifai.errors import UserError
-from clarifai.utils.logging import logger
+from clarifai.utils.logs import logger
 
 
 class ComputeCluster(Lister, BaseClient):
