+++ clarifai/client/deployment.py
@@ -2,7 +2,7 @@ from clarifai_grpc.grpc.api import resources_pb2
 
 from clarifai.client.base import BaseClient
 from clarifai.client.lister import Lister
-from clarifai.utils.logging import logger
+from clarifai.utils.logs import logger
 
 
 class Deployment(Lister, BaseClient):
