+++ clarifai/client/module.py
@@ -6,7 +6,7 @@ from clarifai.client.base import BaseClient
 from clarifai.client.lister import Lister
 from clarifai.errors import UserError
 from clarifai.urls.helper import ClarifaiUrlHelper
-from clarifai.utils.logging import logger
+from clarifai.utils.logs import logger
 
 
 class Module(Lister, BaseClient):
