+++ clarifai/client/nodepool.py
@@ -10,7 +10,7 @@ from clarifai.client.base import BaseClient
 from clarifai.client.deployment import Deployment
 from clarifai.client.lister import Lister
 from clarifai.errors import UserError
-from clarifai.utils.logging import logger
+from clarifai.utils.logs import logger
 
 
 class Nodepool(Lister, BaseClient):
