+++ clarifai/runners/server.py
@@ -13,7 +13,7 @@ from clarifai_protocol.utils.grpc_server import GRPCServer
 from clarifai.runners.models.model_builder import ModelBuilder
 from clarifai.runners.models.model_runner import ModelRunner
 from clarifai.runners.models.model_servicer import ModelServicer
-from clarifai.utils.logging import logger
+from clarifai.utils.logs import logger
 
 
 def main():
